[RHP-103] refactor: add type conversion utility and update Redis client configuration
- plugins/__init__.py: import and expose convert_to_type function
- plugins/convert_types/__init__.py: implement convert_to_type function for string type conversion
- plugins/redis_client_bot/config.py: remove inline convert_to_type function and import from convert_types
