+++ plugins/__init__.py
@@ -10,6 +10,7 @@ Example:
 
 """
 
+from .convert_types import convert_to_type
 from .redis_client_bot import Redis
 
-__all__ = ["Redis"]
+__all__ = ["Redis", "convert_to_type"]
