+++ plugins/redis_client_bot/config.py
@@ -2,36 +2,16 @@
 
 from __future__ import annotations
 
-import json
-from contextlib import suppress
 from typing import Optional, Union
 
 from redis import CredentialProvider
 from redis.cache import CacheConfig, CacheInterface
 
+from plugins import convert_to_type
+
 ValueTypes = Union[str, int, float, bool, CredentialProvider, CacheConfig, CacheInterface, dict, list]
 
 TypesArg = dict[str, ValueTypes]
-
-
-def convert_to_type(val: str) -> ValueTypes:
-    """Convert a string value to the appropriate type."""
-    with suppress(ValueError):
-        return int(val)
-
-    with suppress(ValueError):
-        return float(val)
-
-    if val.lower() == "true":
-        return True
-
-    if val.lower() == "false":
-        return False
-
-    with suppress(json.JSONDecodeError, TypeError, Exception):
-        return json.loads(val)
-
-    return val
 
 
 class ConfigRedis:
